TMO_SHELL DFU: Fixes for dfu outlined below
There are fixes here to get the user defined base url working again.
There are also file name changes for the modem.  The file name that
DFU uses to program the modem after download the updates was changed.
The SHA for the modem ws cleared to zero as it isn't being used.

Signed-off-by: Kim Mansfield <kim.mansfield@2t-mobile.com