Button press callback sample module
This commit pushes a new sample module that demonstrates to use a
callback mechanism when user presses the button.

The callback should be implemented in the user application code, the
callback's 3 parameters would represent the timestamp when the button
was pressed (in millisecs), the duration of how long the button was held
(also in millisecs) and the type of the button press as an enum value:
normal or timeout.

The 'timeout' button type indicates that the user had held the button
down for longer than a pre-configured duration (e.g, to factory reset
the board).

This module also has implementation that the user can de-init and
re-configure the tiemout duration and/or the callback function address.

Signed-off-by: YHusain1 <yusuf.husainy1@t-mobile.com>